+++ src/kcd2_init.cpp
@@ -7,6 +7,7 @@
 
 #include <config/config.hpp>
 #include <gui/gui.hpp>
+#include <gui/ingame_chat.hpp>
 #include <gui/native_multiplayer_menu.hpp>
 #include <gui/renderer.hpp>
 #include <kcse/client_proxy.hpp>
@@ -368,7 +369,8 @@ namespace big
 
 	static void hook_PostInputEvent(char **a1, __int64 a2, __int64 a3, __int64 a4)
 	{
-		if (native_multiplayer_menu::blocks_game_input())
+		if (native_multiplayer_menu::blocks_game_input()
+		    || ingame_chat::blocks_game_input())
 		{
 			return;
 		}